listener: Add new 'newConnection()' callback
Add new 'newConnection()' callback to ListenerFilterCallbacks, which
allows the listener filter to spawn new connections without
terminating the listener filter execution. The main purpose of this is
to allow multiplexing multiple connections on top of a single accepted
socket. Such multiplexing requires a new extension transport socket
implementation, so this commit is only the minimal core Envoy
functionality that is neede to allow such extensions to work.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>